requirements.txt
requests==2.31.0
betterproto==2.0.0b6
websocket-client==1.7.0
PyExecJS==1.5.1
mini_racer==0.12.4

# Flask Web Framework
Flask==3.0.0
Flask-SocketIO==5.3.6
python-socketio==5.10.0

# Database
SQLAlchemy==2.0.23
pymysql==1.1.0

# Task Scheduling
APScheduler==3.10.4

# Additional dependencies
Werkzeug==3.0.1
loguru==0.7.3
python-dotenv==1.0.1
